Big rename of files to be better inline with dart conventions
