+++ lib/gsn/ABI/IRelayHub.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:rly_network_flutter_sdk/gsnClient/ABI/IRelayHubData.dart';
+import 'package:rly_network_flutter_sdk/gsn/ABI/IRelayHubData.dart';
 import 'package:web3dart/web3dart.dart';
 
 DeployedContract relayHubContract(String contractAddress) {
