+++ lib/gsn/EIP712/meta_transactions.dart
@@ -4,8 +4,8 @@ import 'dart:typed_data';
 import 'package:eth_sig_util/eth_sig_util.dart';
 import 'package:eth_sig_util/util/utils.dart';
 import 'package:rly_network_flutter_sdk/contracts/erc20.dart';
-import 'package:rly_network_flutter_sdk/gsnClient/gsnTxHelpers.dart';
-import 'package:rly_network_flutter_sdk/gsnClient/utils.dart';
+import 'package:rly_network_flutter_sdk/gsn/gsn_tx_helpers.dart';
+import 'package:rly_network_flutter_sdk/gsn/utils.dart';
 import 'package:web3dart/web3dart.dart' as web3;
 import 'package:convert/convert.dart';
 import '../../wallet.dart';
