+++ lib/gsn/EIP712/permit_transaction.dart
@@ -3,7 +3,7 @@ import 'dart:convert';
 import 'package:eth_sig_util/eth_sig_util.dart';
 import 'package:eth_sig_util/util/utils.dart';
 import 'package:web3dart/web3dart.dart' as web3;
-import 'package:rly_network_flutter_sdk/gsnClient/gsnTxHelpers.dart';
+import 'package:rly_network_flutter_sdk/gsn/gsn_tx_helpers.dart';
 
 import '../../wallet.dart';
 import '../../contracts/erc20.dart';
