+++ lib/gsn/EIP712/relay_request.dart
@@ -1,0 +1,16 @@
+import 'forward_request.dart';
+import 'relay_data.dart';
+
+class RelayRequest {
+  ForwardRequest request;
+  RelayData relayData;
+
+  RelayRequest({required this.request, required this.relayData});
+  List<dynamic> toJson() {
+    return [request.toJson(), relayData.toJson()];
+  }
+
+  Map<String, dynamic> toMap() {
+    return {"request": request.toMap(), "relayData": relayData.toMap()};
+  }
+}
